security oauth2 com github, melhorias significativas no front e no backend

+++ src/main/java/com/muriloscorp/codesv/controller/SnippetController.java
@@ -1,14 +1,21 @@
 package com.muriloscorp.codesv.controller;
 
 import com.muriloscorp.codesv.dto.SnippetRequest;
+import com.muriloscorp.codesv.exception.UserNotFoundException;
 import com.muriloscorp.codesv.model.CodeSnippet;
+import com.muriloscorp.codesv.model.User;
+import com.muriloscorp.codesv.repository.UserRepository;
 import com.muriloscorp.codesv.service.SnippetService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.UUID;
@@ -18,21 +25,61 @@ import java.util.UUID;
 public class SnippetController {
 
     private final SnippetService snippetService;
+    private final UserRepository userRepository;
 
-    public SnippetController(SnippetService snippetService) {
+    public SnippetController(SnippetService snippetService, UserRepository userRepository) {
         this.snippetService = snippetService;
+        this.userRepository = userRepository;
+    }
+
+    public boolean isOwner(OAuth2User principal, CodeSnippet snippet) {
+        if (principal == null || snippet == null) return false;
+
+        Long githubId = ((Number) principal.getAttributes().get("id")).longValue();
+        return snippet.getAuthor().getGithubId().equals(githubId);
     }
 
     @GetMapping
-    public String listAll(Model model) {
-        model.addAttribute("snippets", snippetService.findAll());
+    public String listAll(
+            Model model, @AuthenticationPrincipal OAuth2User principal,
+            @RequestParam(name = "filter", required = false, defaultValue = "my") String filter
+    ) {
+        if (principal == null) {
+            model.addAttribute("snippets", snippetService.findAll());
+            model.addAttribute("pageTitle", "Explorar Comunidade");
+            model.addAttribute("activerFilter", "all");
+            return "snippet-list";
+        }
+
+        Long githubId = ((Number) principal.getAttributes().get("id")).longValue();
+        User user = userRepository.findByGithubId(githubId)
+                .orElseThrow(() -> new UserNotFoundException("Erro de sessão."));
+
+        if ("all".equals(filter)) {
+            model.addAttribute("snippets", snippetService.findAll());
+            model.addAttribute("pageTitle", "Explorar Comunidade");
+            model.addAttribute("activeFilter", "all");
+            model.addAttribute("isPersonalView", false);
+        } else {
+            model.addAttribute("snippets", snippetService.findByAuthorId(user.getId()));
+            model.addAttribute("pageTitle", "Meus Snippets");
+            model.addAttribute("activeFilter", "my");
+            model.addAttribute("isPersonalView", true);
+        }
+
         return "snippet-list";
     }
 
     @GetMapping("/{id}")
-    public String viewSnippet(@PathVariable UUID id, Model model) {
+    public String viewSnippet(
+            @PathVariable UUID id, Model model,
+            @AuthenticationPrincipal OAuth2User principal
+    ) {
         CodeSnippet snippet = snippetService.findById(id);
+        boolean isOwner = isOwner(principal, snippet);
+
         model.addAttribute("snippet", snippet);
+        model.addAttribute("isOwner", isOwner);
         return "snippet-view";
     }
 
@@ -44,11 +91,18 @@ public class SnippetController {
 
     @PostMapping("/save")
     public String saveSnippet(@Valid @ModelAttribute("snippetForm") SnippetRequest request,
-                              BindingResult result, RedirectAttributes redirect) {
+                              BindingResult result, RedirectAttributes redirect,
+                              @AuthenticationPrincipal OAuth2User principal
+                              ) {
         if (result.hasErrors()) {
             return "snippet-form";
         }
-        snippetService.save(request);
+
+        Long githubId = ((Number) principal.getAttributes().get("id")).longValue();
+        User author = userRepository.findByGithubId(githubId)
+                        .orElseThrow(() -> new UserNotFoundException("Usuário não encontrado."));
+
+        snippetService.save(request, author);
         redirect.addFlashAttribute("success", "Snippet criado com sucesso!");
 
         return "redirect:/snippets";
@@ -70,11 +124,17 @@ public class SnippetController {
             @Valid @ModelAttribute("snippetForm") SnippetRequest request,
             BindingResult result,
             Model model,
-            RedirectAttributes redirect
+            RedirectAttributes redirect,
+            @AuthenticationPrincipal OAuth2User principal
     ) {
         if (result.hasErrors()) {
             model.addAttribute("snippetId", id);
             return "snippet-form";
+        }
+
+        CodeSnippet snippet = snippetService.findById(id);
+        if (!isOwner(principal, snippet)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acesso Negado.");
         }
 
         snippetService.update(id, request);
@@ -84,9 +144,20 @@ public class SnippetController {
     }
 
     @PostMapping("/{id}/delete")
-    public String deleteSnippet(@PathVariable UUID id, RedirectAttributes redirect) {
+    public String deleteSnippet(
+            @PathVariable UUID id, RedirectAttributes redirect,
+            @AuthenticationPrincipal OAuth2User principal
+    ) {
+        CodeSnippet snippet = snippetService.findById(id);
+        if (!isOwner(principal, snippet)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Você não tem permissão para excluir esse snippet.");
+        }
+
         snippetService.delete(id);
         redirect.addFlashAttribute("success", "Snippet excluído com sucesso!");
         return "redirect:/snippets";
     }
+
+    @GetMapping("/{id}/download")
+    public String downloadSnippet(@PathVariable UUID id) {}
 }
